MMT-3551: Removed the middle man class aclConcept.js,removed its test file, modified README
