+++ src/cmr/concepts/__tests__/acl.test.js
@@ -53,4 +53,36 @@ describe('Acl', () => {
       }
     ])
   })
+
+  test('fetchAcl calls aclQuery with correct parameters when jsonKeys are present', () => {
+    const acl = new Acl({ 'Header-Key': 'header-value' }, { jsonKeys: ['key1', 'key2'] })
+    acl.jsonItemCount = 1
+
+    const aclQuerySpy = jest.spyOn(acl, 'fetchAcl').mockImplementation(() => Promise.resolve())
+
+    acl.fetch({
+      key1: 'value1',
+      key2: 'value2'
+    })
+
+    expect(aclQuerySpy).toHaveBeenCalledWith({
+      key1: 'value1',
+      key2: 'value2'
+    }, ['key1', 'key2'], { 'Header-Key': 'header-value' })
+  })
+
+  test('fetchAcl does not call aclQuery when jsonKeys are not present', () => {
+    const acl = new Acl({ 'Header-Key': 'header-value' }, { jsonKeys: [] })
+    acl.jsonItemCount = 42
+
+    const aclQuerySpy = jest.spyOn(acl, 'fetchAcl').mockImplementation(() => Promise.resolve())
+
+    acl.fetch({
+      key1: 'value1',
+      key2: 'value2'
+    })
+
+    // Expect that aclQuery is not called when jsonKeys are not present
+    expect(aclQuerySpy).not.toHaveBeenCalled()
+  })
 })
