+++ src/cmr/concepts/acl.js
@@ -1,6 +1,11 @@
-import AclConcept from './aclConcept'
+import snakeCaseKeys from 'snakecase-keys'
+import { pick } from 'lodash'
+import { aclQuery } from '../../utils/aclQuery'
+import { mergeParams } from '../../utils/mergeParams'
+import { downcaseKeys } from '../../utils/downcaseKeys'
+import Concept from './concept'
 
-export default class Acl extends AclConcept {
+export default class Acl extends Concept {
   /**
    * Instantiates an ACL object from the CMR API
    * @param {Object} headers HTTP headers provided by the query
@@ -23,13 +28,48 @@ export default class Acl extends AclConcept {
     return items
   }
 
+  fetch(searchParams) {
+    const params = mergeParams(searchParams)
+
+    // Default an array to hold the promises we need to make depending on the requested fields
+    const promises = []
+
+    const { jsonKeys } = this.requestInfo
+
+    if (jsonKeys.length > 0) {
+      const jsonHeaders = {
+        ...this.headers
+      }
+      promises.push(
+        this.fetchAcl(params, jsonKeys, jsonHeaders)
+      )
+    } else {
+      // Push a null promise to the array so that the umm promise always exists as
+      // the second element of the promise array
+      promises.push(
+        // eslint-disable-next-line no-promise-executor-return
+        new Promise((resolve) => resolve(null))
+      )
+    }
+
+    this.response = Promise.all(promises)
+  }
+
   /**
-   * Query the CMR ACL API endpoint to retrieve requested data
+   * Query the CMR JSON API endpoint to retrieve requested data
    * @param {Object} searchParams Parameters provided by the query
-   * @param {Array} ummKeys Keys requested by the query
-   * @param {Object} headers Headers requested by the query
+   * @param {Array} requestedKeys Keys requested by the query
+   * @param {Object} providedHeaders Headers requested by the query
    */
-  fetchAcl(searchParams, ummKeys, headers) {
-    return super.fetchAcl(searchParams, ummKeys, headers)
+  fetchAcl(searchParams, requestedKeys, providedHeaders) {
+    this.logKeyRequest(requestedKeys, 'json')
+
+    // Construct the promise that will request data from the json endpoint
+    return aclQuery({
+      conceptType: this.getConceptType(),
+      params: pick(snakeCaseKeys(searchParams), this.getPermittedJsonSearchParams()),
+      nonIndexedKeys: this.getNonIndexedKeys(),
+      headers: providedHeaders
+    })
   }
 }
